+++ src/c/lib.h
@@ -7,7 +7,7 @@ char *strcpy(char *dest, const char *src);
 int strcmp(const char *s1, const char *s2);
 const char *strchr(const char *s, int c);
 
-void to_string(int i, char* s, size_t base);
+void to_string(long i, char* s, long base);
 
 // Reasoning in lib.c
 // char *strdup(const char *s);
